Added memo_store_local_loader.dart.
