+++ lib/memo_store_loader.dart
@@ -21,9 +21,6 @@
  */
 
 import "dart:convert";
-import 'dart:io';
-
-import 'package:path_provider/path_provider.dart';
 
 import 'memo.dart';
 import 'memo_store.dart';
@@ -67,32 +64,5 @@ class MemoStoreLoader {
       memo.lastMergedRevision = deserializedMemo['lastMergedRevision'];
       _memoStore.addMemo(memo);
     }
-  }
-}
-
-class MemoStoreLocalLoader extends MemoStoreLoader {
-  final String _path;
-
-  /// Creates a memo store loader.
-  MemoStoreLocalLoader(MemoStore memoStore, this._path) : super(memoStore);
-
-  /// Executes this memo store loader.
-  Future<void> execute() async {
-    final file = File(_path);
-    final string = await file.readAsString();
-    deserialize(string);
-  }
-
-  /// Creates a memo store loader from file name.
-  static Future<MemoStoreLocalLoader> fromFileName(
-      MemoStore memoStore, String fileName) async {
-    final applicationDocumentsDirectory =
-        await getApplicationDocumentsDirectory();
-    var path = applicationDocumentsDirectory.path;
-    print('path: ${path}\n');
-    path = path + Platform.pathSeparator + fileName;
-    print('path: ${path}\n');
-
-    return MemoStoreLocalLoader(memoStore, path);
   }
 }
